Merge settings files into common

+++ base_archive/cleanup.sh
@@ -8,26 +8,25 @@
 
 # include working environment
 . ../common/init.sh
-. ./settings.sh
 
 # clean all the buildid-caches created within this test
 touch $CURRENT_TEST_DIR/BUILDIDDIRS
 while read line; do
 	BUILDIDDIR="$line"
-	if [[ " ${BUILDIDDIR}/*" =~ ' /*' ]] || [[ "$BUILDIDDIR" =~ /$ ]]; then
+	if [[ " ${BUILDIDDIR}/*" == *' /*'* ]] || [[ "$BUILDIDDIR" =~ /$ ]]; then
 		true # skipping deletion
 	else
 		rm -rf $BUILDIDDIR/.b*
-		rm -rf $BUILDIDDIR/*
+		rm -rf ${BUILDIDDIR:?}/*
 		rmdir $BUILDIDDIR 2> /dev/null
 	fi
 done < $CURRENT_TEST_DIR/BUILDIDDIRS
 rm -f $CURRENT_TEST_DIR/BUILDIDDIRS
 
 if [ ! -n "$PERFSUITE_RUN_DIR" ]; then
-	find . -name \*.log | xargs -r rm
-	find . -name \*.err | xargs -r rm
-	find . -name \*.list | xargs -r rm
+	find . -name \*.log -print0 | xargs -0 -r rm
+	find . -name \*.err -print0 | xargs -0 -r rm
+	find . -name \*.list -print0 | xargs -0 -r rm
 	rm -f perf.data*
 fi
